This solves issue #29614 as an enhancement

now `repeat(ones(2,2), 2,2,2)` won't raise `BoundsError` but instead return a `Array{T,3}` as intuition expects. 

Benchmark:
```julia
julia> a = ones(100,100,100);
# before this patch
julia> @btime repeat(a, 2,2,2);
  81.971 ms (40 allocations: 114.44 MiB)
# after this patch
julia> @btime repeat(a, 2,2,2);
  82.136 ms (40 allocations: 114.44 MiB)
```