@johnnychen94 2018-10-17
Thanks for the reviewing @mbauman and sorry for the back-and-forth commits on the same code.

I am not very familiar with the index mechanisms of Julia Arrays; I thought `1` is equivalent to `OneTo(1)` in this case, but the test says no.

I was even thinking of using 
```julia
R[axes(A)...,ones(Int, length(shape) - length(inner))...] = A
```
to replace
```julia
idxs = (axes(A)..., ntuple(n->OneTo(1), length(shape) - length(inner))...)
R[idxs...] = A
```
But not very sure if this works. (Actually I don't know how to fully test it in my own machine)