@mbauman 2018-10-17
The best way to test changes like this locally is to use Revise.jl, but sometimes it's quickest/easiest to [just `@eval` the code in the REPL](https://gist.github.com/mbauman/e5d35b780aed46d6ab69c79a31484446).  Just type `@eval Base` (where `Base` is the module in which the original code was defined) and then paste the new definition.

Sorry for the misdirection on the colons — of course we just want to be assigning into the _first_ block.  Using `1` instead of `OneTo(1)` does indeed behave the same in `setindex!` and does indeed work… you just got bit by a spurious network interruption in an unrelated test on CI.  I do find `1` to be a bit simpler and clearer in intent, but it's not a big deal.  More importantly, though, I think you want to just use `ndims(R) - ndims(A)` instead of the lengths of inner and shape.  Could you add this test case?

```julia
repeat(ones(2,2), inner=(1, 1, 1), outer=(2, 2, 2))
```

With respect to your question about `ones`: the challenge is doing these index manipulations in a way that is type-stable — that is, in a way where Julia can infer the length at compile time.  Splatting an array like `ones(…)` into a tuple is definitely not what you want in that case as it allocates an array and Julia doesn't store array lengths in the type system.  Using `ntuple` is better — especially if we can use an inferable constant for the second argument… but that subtraction isn't inferable.  It looks like this whole function could use a bit of hand optimizing… so making this better here can be future work.  Also good future work could be making this work in cases where there is an inner repeat.