@johnnychen94 2018-10-18
Thanks for this explanation, this helps me a lot.

With regard to optimization,  I did some relative benchmark tests as a reference (kind of interesting to me). This benchmark says, at least, we can optimize `repeat` for some special cases.

```julia
# Baseline
# without this patch
julia> @btime repeat(ones(100,100), 20, 20);
  8.828 ms (4 allocations: 30.59 MiB)

# without this patch (slightly faster than Baseline)
julia> @btime reshape(repeat(ones(100,100),20,20),2000,2000,1,1,1);
  8.590 ms (6 allocations: 30.59 MiB)

# without this patch
julia> @btime repeat(reshape(ones(100,100),100,100,1,1,1),20,20,1,1,1);
66.962 ms (3625 allocations: 61.17 MiB)

# with this patch
julia> @btime repeat(ones(100,100), 20, 20,1,1,1);
70.049 ms (3625 allocations: 61.17 MiB)
```

I'll tell you when I have a solution to support this feature without loss much performance. (I have some other jobs to do, so no sooner than Nov)
But still, I think it's okay to merge this now and recreate a new PR when we find a way to optimize this.