@johnnychen94 2018-10-20
ping @mbauman I think this is good to merge. Can you have a quick look?