@mbauman 2018-10-23
Yes, you're exactly right and this did indeed work previously. It had the somewhat strange behavior that the trailing index was always allowed to "linearly span" (working through memory order for plain old `Array`) over any omitted dimensions.  This ran us into trouble in a number of situations, including consistency with offset arrays, the meaning of the `end` keyword, and expectations of bounds errors. For more details see: #14470, #20079, and #21750.  We now have the following rules:

* Indexing with one index is linear indexing
* Indexing with less than `ndims` indices is only supported if the dimensions omitted are all "singleton" (e.g., length 1) 
* Indexing with more than `ndims` indices is only supported if the extra dimensions provided are all selecting the first index (e.g., 1)

In other words, you've found and cleaned up some of my own dirty laundry — thank you!